Moved SLSP protocol files
