+++ src/CTDataProvider.ts
@@ -2,7 +2,7 @@
 
 import { Event, EventEmitter, ExtensionContext, TreeDataProvider, TreeItem, TreeItemCollapsibleState } from 'vscode';
 import { CTTreeView } from './CTTreeView';
-import { NumberRange, VerdictKind } from './protocol/combinatorialTesting.slsp';
+import { NumberRange, VerdictKind } from './protocol/slsp/combinatorialTesting';
 import { Icons } from './Icons'
 
 export class CTDataProvider implements TreeDataProvider<TestViewElement> {
