+++ src/protocol/slsp/proofObligationGeneration.ts
